requirements.txt
obspy
scipy
pandas
numpy
openpyxl
plotly
streamlit-extras
streamlit-option-menu
XlsxWriter
parse
